Добавила страницу 404 в шаблон
